+++ proptool/checks/base/check.py
@@ -46,8 +46,7 @@ class Check(ABC):
         return Check.DEFAULT_CHECK_CONFIG
 
     def load_config_ini(self, config: Dict, parser: ConfigParser, config_section: str) -> None:
-        options = list(self.get_default_config().keys())
-        self._set_config_option(config, parser, config_section, options)
+        self._set_config_option(config, parser, config_section, self.config)
 
     def _set_config_option(self, config: Dict, parser: ConfigParser, config_section: str, options: Union[List[str], str]) -> None:
         if isinstance(options, str):