+++ proptool/config/config.py
@@ -47,6 +47,7 @@ class Config(object):
 
         self.update: bool = False
         self.create: bool = False
+        self.write_reference: bool = False
 
         self.files: List[str] = []
         self.languages: List[str] = []