+++ proptool/const.py
@@ -20,3 +20,11 @@ class Const(object):
         'The *.properties file checker and syncing tool.',
         f'{APP_URL}',
     ]
+
+    class RC(object):  # noqa: WPS431
+        """
+        Application return codes.
+        """
+
+        OK: int = 0
+        TRANSLATION_SYNTAX_ERROR = 200